+++ zellij-server/src/output/mod.rs
@@ -1,15 +1,14 @@
-//! `Tab`s holds multiple panes. It tracks their coordinates (x/y) and size,
-//! as well as how they should be resized
-
 use zellij_utils::position::{Column, Line};
 use zellij_utils::{position::Position, serde, zellij_tile};
+use std::collections::VecDeque;
 
 use crate::ui::pane_boundaries_frame::FrameParams;
-// use pane_grid::{split, FloatingPaneGrid, PaneGrid};
+use crate::panes::selection::Selection;
+use crate::panes::Row;
 
 use crate::{
     os_input_output::ServerOsApi,
-    panes::{PaneId, PluginPane, TerminalPane, TerminalCharacter, EMPTY_TERMINAL_CHARACTER, CharacterChunk, LinkHandler},
+    panes::{PaneId, PluginPane, TerminalPane, TerminalCharacter, EMPTY_TERMINAL_CHARACTER, LinkHandler},
     pty::{ClientOrTabIndex, PtyInstruction, VteBytes},
     thread_bus::ThreadSenders,
     ui::boundaries::Boundaries,
@@ -17,7 +16,7 @@ use crate::{
     wasm_vm::PluginInstruction,
     ClientId, ServerInstruction,
     panes::terminal_character::{
-        CharacterStyles, CursorShape,
+        CharacterStyles, CursorShape, AnsiCode,
     },
 };
 use serde::{Deserialize, Serialize};
@@ -40,21 +39,6 @@ use zellij_utils::{
     },
     pane_size::{Offset, PaneGeom, Size, Viewport},
 };
-
-macro_rules! resize_pty {
-    ($pane:expr, $os_input:expr) => {
-        if let PaneId::Terminal(ref pid) = $pane.pid() {
-            // FIXME: This `set_terminal_size_using_fd` call would be best in
-            // `TerminalPane::reflow_lines`
-            $os_input.set_terminal_size_using_fd(
-                *pid,
-                $pane.get_content_columns() as u16,
-                $pane.get_content_rows() as u16,
-            );
-        }
-    };
-}
-
 
 // FIXME: This should be replaced by `RESIZE_PERCENT` at some point
 const MIN_TERMINAL_HEIGHT: usize = 5;
@@ -241,5 +225,205 @@ impl FloatingPanesStack {
             }
         }
         visible_chunks
+    }
+}
+
+#[derive(Debug, Clone, Default)]
+pub struct CharacterChunk {
+    pub terminal_characters: Vec<TerminalCharacter>,
+    pub x: usize,
+    pub y: usize,
+    pub changed_colors: Option<[Option<AnsiCode>; 256]>,
+    selection_and_background_color: Option<(Selection, AnsiCode)>,
+}
+
+impl CharacterChunk {
+    pub fn new(terminal_characters: Vec<TerminalCharacter>, x: usize, y: usize) -> Self {
+        CharacterChunk {
+            terminal_characters,
+            x,
+            y,
+            ..Default::default()
+        }
+    }
+    pub fn add_selection_and_background(&mut self, selection: Selection, background_color: AnsiCode, offset_x: usize, offset_y: usize) {
+        self.selection_and_background_color = Some((selection.offset(offset_x, offset_y), background_color));
+    }
+    pub fn selection_and_background_color(&self) -> Option<(Selection, AnsiCode)> {
+        self.selection_and_background_color
+    }
+    pub fn add_changed_colors(&mut self, changed_colors: Option<[Option<AnsiCode>; 256]>) {
+        self.changed_colors = changed_colors;
+    }
+    pub fn changed_colors(&self) -> Option<[Option<AnsiCode>; 256]> {
+        self.changed_colors
+    }
+    pub fn width(&self) -> usize {
+        let mut width = 0;
+        for t_character in &self.terminal_characters {
+            width += t_character.width
+        }
+        width
+    }
+    pub fn drain_by_width(&mut self, x: usize) -> impl Iterator<Item=TerminalCharacter> {
+        let mut drained_part: VecDeque<TerminalCharacter> = VecDeque::new();
+        let mut drained_part_len = 0;
+        loop {
+            if self.terminal_characters.is_empty() {
+                break;
+            }
+            let next_character = self.terminal_characters.remove(0); // TODO: consider copying self.terminal_characters into a VecDeque to make this process faster?
+            if drained_part_len + next_character.width <= x {
+                drained_part.push_back(next_character);
+                drained_part_len += next_character.width;
+            } else {
+                if drained_part_len == x {
+                    self.terminal_characters.insert(0, next_character); // put it back
+                } else if next_character.width > 1 {
+                    for _ in 1..next_character.width {
+                        self.terminal_characters.insert(0, EMPTY_TERMINAL_CHARACTER);
+                        drained_part.push_back(EMPTY_TERMINAL_CHARACTER);
+                    }
+                }
+                break;
+            }
+        }
+        drained_part.into_iter()
+    }
+    pub fn retain_by_width(&mut self, x: usize) {
+        let part_to_retain = self.drain_by_width(x); // TODO: better
+        self.terminal_characters = part_to_retain.collect();
+    }
+    pub fn cut_middle_out(&mut self, middle_start: usize, middle_end: usize) -> (Vec<TerminalCharacter>, Vec<TerminalCharacter>) {
+        let mut absolute_middle_start_index = None;
+        let mut absolute_middle_end_index = None;
+        let mut current_x = 0;
+        let mut pad_left_end_by = 0;
+        let mut pad_right_start_by = 0;
+        for (absolute_index, t_character) in self.terminal_characters.iter().enumerate() {
+            current_x += t_character.width;
+            if current_x >= middle_start && absolute_middle_start_index.is_none() {
+                if current_x > middle_start {
+                    pad_left_end_by = current_x - middle_start;
+                    absolute_middle_start_index = Some(absolute_index);
+                } else {
+                    absolute_middle_start_index = Some(absolute_index + 1);
+                }
+            }
+            if current_x >= middle_end && absolute_middle_end_index.is_none() {
+                absolute_middle_end_index = Some(absolute_index + 1);
+                if current_x > middle_end {
+                    pad_right_start_by = current_x - middle_end;
+                }
+            }
+        }
+        let absolute_middle_start_index = absolute_middle_start_index.unwrap();
+        let absolute_middle_end_index = absolute_middle_end_index.unwrap();
+        let mut terminal_characters: Vec<TerminalCharacter> = self.terminal_characters.drain(..).collect();
+        let mut characters_on_the_right: Vec<TerminalCharacter> = terminal_characters.drain(absolute_middle_end_index..).collect();
+        let mut characters_on_the_left: Vec<TerminalCharacter>  = terminal_characters.drain(..absolute_middle_start_index).collect();
+        if pad_left_end_by > 0 {
+            for _ in 0..pad_left_end_by {
+                characters_on_the_left.push(EMPTY_TERMINAL_CHARACTER);
+            }
+        }
+        if pad_right_start_by > 0 {
+            for _ in 0..pad_right_start_by {
+                characters_on_the_right.insert(0, EMPTY_TERMINAL_CHARACTER);
+            }
+        }
+        (characters_on_the_left, characters_on_the_right)
+    }
+}
+
+#[derive(Clone, Debug)]
+pub struct OutputBuffer {
+    changed_lines: Vec<usize>, // line index
+    should_update_all_lines: bool,
+}
+
+impl Default for OutputBuffer {
+    fn default() -> Self {
+        OutputBuffer {
+            changed_lines: vec![],
+            should_update_all_lines: true, // first time we should do a full render
+        }
+    }
+}
+
+impl OutputBuffer {
+    pub fn update_line(&mut self, line_index: usize) {
+        if !self.should_update_all_lines {
+            self.changed_lines.push(line_index);
+        }
+    }
+    pub fn update_all_lines(&mut self) {
+        self.clear();
+        self.should_update_all_lines = true;
+    }
+    pub fn clear(&mut self) {
+        self.changed_lines.clear();
+        self.should_update_all_lines = false;
+    }
+    pub fn changed_chunks_in_viewport(
+        &self,
+        viewport: &[Row],
+        viewport_width: usize,
+        viewport_height: usize,
+        x_offset: usize,
+        y_offset: usize,
+    ) -> Vec<CharacterChunk> {
+        if self.should_update_all_lines {
+            let mut changed_chunks = Vec::with_capacity(viewport.len());
+            for line_index in 0..viewport_height {
+                let terminal_characters =
+                    self.extract_line_from_viewport(line_index, viewport, viewport_width);
+                let x = 0 + x_offset; // right now we only buffer full lines as this doesn't seem to have a huge impact on performance, but the infra is here if we want to change this
+                let y = line_index + y_offset;
+                changed_chunks.push(CharacterChunk::new(terminal_characters, x, y));
+            }
+            changed_chunks
+        } else {
+            let mut line_changes = self.changed_lines.to_vec();
+            line_changes.sort_unstable();
+            line_changes.dedup();
+            let mut changed_chunks = Vec::with_capacity(line_changes.len());
+            for line_index in line_changes {
+                let terminal_characters =
+                    self.extract_line_from_viewport(line_index, viewport, viewport_width);
+                let x = 0 + x_offset;
+                let y = line_index + y_offset;
+                changed_chunks.push(CharacterChunk::new(terminal_characters, x, y));
+            }
+            changed_chunks
+        }
+    }
+    fn extract_characters_from_row(
+        &self,
+        row: &Row,
+        viewport_width: usize,
+    ) -> Vec<TerminalCharacter> {
+        let mut terminal_characters: Vec<TerminalCharacter> = row.columns.iter().copied().collect();
+        // pad row
+        let row_width = row.width();
+        if row_width < viewport_width {
+            let mut padding = vec![EMPTY_TERMINAL_CHARACTER; viewport_width - row_width];
+            terminal_characters.append(&mut padding);
+        }
+        terminal_characters
+    }
+    fn extract_line_from_viewport(
+        &self,
+        line_index: usize,
+        viewport: &[Row],
+        viewport_width: usize,
+    ) -> Vec<TerminalCharacter> {
+        match viewport.get(line_index) {
+            // TODO: iterator?
+            Some(row) => self.extract_characters_from_row(row, viewport_width),
+            None => {
+                vec![EMPTY_TERMINAL_CHARACTER; viewport_width]
+            }
+        }
     }
 }
